+++ src/DataFramesMeta.jl
@@ -6,9 +6,13 @@ using MacroTools
 
 using OrderedCollections: OrderedCollections
 
+@reexport using TableMetadataTools
+
 @reexport using DataFrames
 
 @reexport using Chain
+
+using DataFrames.PrettyTables
 
 # Basics:
 export @with,
@@ -20,7 +24,13 @@ export @with,
        @rtransform, @rselect, @rtransform!, @rselect!,
        @distinct, @rdistinct, @distinct!, @rdistinct!,
        @eachrow, @eachrow!,
+<<<<<<< HEAD
        @byrow, @passmissing, @astable, @kwarg, @when,
+=======
+       @byrow, @passmissing, @astable, @kwarg,
+       @label!, @note!, printlabels, printnotes,
+       @groupby,
+>>>>>>> master
        @based_on, @where # deprecated
 
 const DOLLAR = raw"$"
@@ -30,5 +40,6 @@ include("parsing_astable.jl")
 include("macros.jl")
 include("linqmacro.jl")
 include("eachrow.jl")
+include("metadata.jl")
 
 end # module