+++ test/runtests.jl
@@ -16,7 +16,8 @@ my_tests = ["dataframes.jl",
             "astable.jl",
             "astable_flag.jl",
             "passmissing.jl",
-            "when.jl"]
+            "when.jl",
+            "multicol.jl"]
 
 println("Running tests:")