+++ NEWS.md
@@ -1,3 +1,15 @@
+# DataFramesMeta v0.15.1 Release notes
+
+* Deleted an errant `print` statement introduced in [#382](https://github.com/JuliaData/DataFramesMeta.jl/pull/382) ([#389](https://github.com/JuliaData/DataFramesMeta.jl/pull/389))
+
+# DataFramesMeta v0.15.0 Release notes
+
+* Fixed macro hygiene issues ([#384](https://github.com/JuliaData/DataFramesMeta.jl/pull/384))
+* Fixed a bug with broadcasting in `@astable` ([#382](https://github.com/JuliaData/DataFramesMeta.jl/pull/382))
+* Added `@note!` and `@label!`, along with `printlabels` and `printnotes` to make working with metadata easier. ([#377](https://github.com/JuliaData/DataFramesMeta.jl/pull/377))
+* Added support for `Not`, `All`, and `Cols` in `@select`, making it easier to keep or drop many columns at once. ([#372](https://github.com/JuliaData/DataFramesMeta.jl/pull/372))
+* Added the `@groupby` macro, which provides an easier syntax for grouping data frames ([#373](https://github.com/JuliaData/DataFramesMeta.jl/pull/373))
+
 # DataFramesMeta v0.14.1 Release notes
 * Fixed a bug where `ByRow` was not properly identified if DataFramesMeta.jl was `import`ed ([#366](https://github.com/JuliaData/DataFramesMeta.jl/pull/366))